chore: move files to doc path
