+++ .gitignore
@@ -2,3 +2,4 @@ dist
 node_modules
 .vitepress/cache
 .vitepress/dist
+.vitepress/temp
