+++ doc/author/[pkg].md
