+++ doc/plugins/[pkg].md
